+++ lib/sgbm_helper.hpp
@@ -15,13 +15,13 @@ unsigned long long int * out_arr,
 int rows, int cols, int csize);
 
 
-inline unsigned long long int hamming_dist(unsigned long long a,
+inline float hamming_dist(unsigned long long a,
 unsigned long long b);
 
 
 extern "C" void shift_subtract_stack(unsigned long long int * L,
 unsigned long long int * R,
-double * out,
+float * out,
 int rows, int cols, int D);
